backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Pydantic for data validation
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
sqlmodel>=0.0.14

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
bcrypt==4.0.1

# Google OAuth
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Environment variables
python-dotenv>=1.0.0

# AWS S3 (optional, for production)
boto3>=1.34.34

# Email validation
email-validator>=2.0.0
